reddit_kaggle_data_pipeline /requirements.txt
praw
google-cloud-storage
mongodb
pyspark
apache-airflow-providers-apache-spark 
airflow
sodapy
pymongo
pandas
datetime
os